+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/mapping/MappingSpec.scala
@@ -66,6 +66,7 @@ object MappingSpec extends TypeRegistry[MappingSpec] {
     new JsonSubTypes.Type(name = "flatten", value = classOf[FlattenMappingSpec]),
     new JsonSubTypes.Type(name = "groupedAggregate", value = classOf[GroupedAggregateMappingSpec]),
     new JsonSubTypes.Type(name = "historize", value = classOf[HistorizeMappingSpec]),
+    new JsonSubTypes.Type(name = "iterativeSql", value = classOf[IterativeSqlMappingSpec]),
     new JsonSubTypes.Type(name = "join", value = classOf[JoinMappingSpec]),
     new JsonSubTypes.Type(name = "latest", value = classOf[LatestMappingSpec]),
     new JsonSubTypes.Type(name = "mock", value = classOf[MockMappingSpec]),